+++ clients/vscode-hlasmplugin/src/test/suite/hlasmExternalFiles.test.ts
@@ -14,76 +14,14 @@
 
 
 import * as assert from 'assert';
-import { HLASMExternalFilesFtp } from '../../hlasmExternalFilesFtp';
-import { ExternalRequestType, HLASMExternalFiles } from '../../hlasmExternalFiles';
-import { ExtensionContext } from 'vscode';
+import * as crypto from "crypto";
+import { ClientUriDetails, ExternalFilesClient, ExternalRequestType, HLASMExternalFiles } from '../../hlasmExternalFiles';
+import { FileSystem, Uri } from 'vscode';
+import { FileType } from 'vscode';
+import { TextEncoder } from 'util';
+import { deflateSync } from 'zlib';
 
-const extensionContextMock = undefined as any as ExtensionContext;
-
-suite('External files (FTP)', () => {
-
-    test('Dataset parsing', async () => {
-        const ftpClient = new HLASMExternalFilesFtp(extensionContextMock);
-
-        assert.strictEqual(ftpClient.parseArgs('aaa', ExternalRequestType.read_directory), null);
-        assert.strictEqual(ftpClient.parseArgs('/0', ExternalRequestType.read_directory), null);
-        assert.strictEqual(ftpClient.parseArgs('/AAAAAAAAA', ExternalRequestType.read_directory), null);
-        assert.strictEqual(ftpClient.parseArgs('/AAAAAAAA/BBBBBBB', ExternalRequestType.read_directory), null);
-        assert.strictEqual(ftpClient.parseArgs('/AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAA.A', ExternalRequestType.read_directory), null);
-
-        const full_length = ftpClient.parseArgs('/aaaaaaaa.aaaaaaaa.aaaaaaaa.aaaaaaaa.aaaaaaaa', ExternalRequestType.read_directory);
-        assert.ok(full_length);
-        assert.strictEqual(full_length.toString(), 'AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA');
-        assert.strictEqual(full_length.normalizedPath(), '/AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA/');
-    });
-
-    test('Dataset member parsing', async () => {
-        const ftpClient = new HLASMExternalFilesFtp(extensionContextMock);
-
-        assert.strictEqual(ftpClient.parseArgs('aaa', ExternalRequestType.read_file), null);
-        assert.strictEqual(ftpClient.parseArgs('/0', ExternalRequestType.read_file), null);
-        assert.strictEqual(ftpClient.parseArgs('/AAAAAAAAA', ExternalRequestType.read_file), null);
-        assert.strictEqual(ftpClient.parseArgs('/AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAA.A', ExternalRequestType.read_file), null);
-
-        assert.strictEqual(ftpClient.parseArgs('/AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA', ExternalRequestType.read_file), null);
-        assert.strictEqual(ftpClient.parseArgs('/AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA/0', ExternalRequestType.read_file), null);
-
-        assert.strictEqual(ftpClient.parseArgs('/AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA/BBBBBBBBB', ExternalRequestType.read_file), null);
-
-        const full_length = ftpClient.parseArgs('/aaaaaaaa.aaaaaaaa.aaaaaaaa.aaaaaaaa.aaaaaaaa/bbbbbbbb', ExternalRequestType.read_file);
-        assert.ok(full_length);
-        assert.strictEqual(full_length.toString(), 'AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA(BBBBBBBB)');
-        assert.strictEqual(full_length.normalizedPath(), '/AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA.AAAAAAAA/BBBBBBBB');
-    });
-
-    test('Suspend', async () => {
-        const ftpClient = new HLASMExternalFilesFtp(extensionContextMock);
-
-        assert.strictEqual(ftpClient.suspended(), false);
-
-        const p = new Promise<boolean>(resolve => ftpClient.onStateChange(resolve));
-
-        ftpClient.suspend();
-
-        assert.strictEqual(await p, true);
-
-        ftpClient.dispose();
-    });
-
-    test('Resume', async () => {
-        const ftpClient = new HLASMExternalFilesFtp(extensionContextMock);
-        ftpClient.suspend();
-
-        assert.strictEqual(ftpClient.suspended(), true);
-
-        const p = new Promise<boolean>(resolve => ftpClient.onStateChange(resolve));
-
-        ftpClient.resume();
-
-        assert.strictEqual(await p, false);
-
-        ftpClient.dispose();
-    });
+suite('External files', () => {
 
     test('Invalid messages', async () => {
         const ext = new HLASMExternalFiles('test', {
@@ -105,5 +43,140 @@ suite('External files (FTP)', () => {
         assert.deepEqual(await ext.handleRawMessage({ id: 5, op: 'read_file', url: 'unknown:scheme' }), { id: 5, error: { code: -5, msg: 'Invalid request' } });
 
         assert.deepEqual(await ext.handleRawMessage({ id: 5, op: 'read_file', url: 'test:/SERVICE' }), { id: 5, error: { code: -1000, msg: 'No client' } });
+    });
+
+    test('Clear cache', async () => {
+        const cacheUri = Uri.parse('test:cache/');
+
+        let readCounter = 0;
+        let deleteCounter = 0;
+
+        const ext = new HLASMExternalFiles('test', {
+            onNotification: (_, __) => { return { dispose: () => { } }; },
+            sendNotification: (_: any, __: any) => Promise.resolve(),
+        }, {
+            uri: cacheUri,
+            fs: {
+                readDirectory: async (uri: Uri) => {
+                    ++readCounter;
+                    assert.strictEqual(cacheUri.toString(), uri.toString());
+                    return [['A', FileType.File]];
+                },
+                delete: async (uri: Uri, options?: { recursive?: boolean; useTrash?: boolean }) => {
+                    ++deleteCounter;
+                    assert.strictEqual(uri.toString(), Uri.joinPath(cacheUri, 'A').toString())
+                },
+            } as any as FileSystem
+        });
+
+        await ext.clearCache();
+
+        assert.strictEqual(readCounter, 1);
+        assert.strictEqual(deleteCounter, 1);
+
+    });
+    const nameGenerator = (components: string[]) => {
+        return 'v1.' + crypto.createHash('sha256').update(JSON.stringify(components)).digest().toString('hex');
+    };
+    test('Access cached content', async () => {
+        const cacheUri = Uri.parse('test:cache/');
+        const dirResponse = deflateSync(new TextEncoder().encode(JSON.stringify("not_exists")));
+        const dir2Content = ['FILE'];
+        const dir2Response = deflateSync(new TextEncoder().encode(JSON.stringify({ data: dir2Content })));
+        const fileContent = 'file content';
+        const fileResponse = deflateSync(new TextEncoder().encode(JSON.stringify({ data: fileContent })));
+
+        let dirWritten = false;
+        let dir2Written = false;
+        let fileWritten = false;
+
+        const ext = new HLASMExternalFiles('test', {
+            onNotification: (_, __) => { return { dispose: () => { } }; },
+            sendNotification: (_: any, __: any) => Promise.resolve(),
+        }, {
+            uri: cacheUri,
+            fs: {
+                readFile: async (uri: Uri) => {
+                    const filename = uri.path.split('/').pop();
+                    if (filename === nameGenerator(['SERVER', 'TEST', '/DIR']))
+                        return dirResponse;
+                    if (filename === nameGenerator(['SERVER', 'TEST', '/DIR2']))
+                        return dir2Response;
+                    if (filename === nameGenerator(['SERVER', 'TEST', '/DIR2/FILE']))
+                        return fileResponse;
+                    assert.ok(false);
+                },
+                writeFile: async (uri: Uri, content: Uint8Array) => {
+                    const filename = uri.path.split('/').pop();
+                    if (filename === nameGenerator(['SERVER', 'TEST', '/DIR'])) {
+                        dirWritten = true;
+                        assert.deepStrictEqual(content, dirResponse);
+                        return;
+                    }
+                    if (filename === nameGenerator(['SERVER', 'TEST', '/DIR2'])) {
+                        dir2Written = true;
+                        assert.deepStrictEqual(content, dir2Response);
+                        return;
+                    }
+                    if (filename === nameGenerator(['SERVER', 'TEST', '/DIR2/FILE'])) {
+                        fileWritten = true;
+                        assert.deepStrictEqual(content, fileResponse);
+                        return;
+                    }
+                    assert.ok(false);
+                },
+            } as any as FileSystem
+        });
+
+        ext.setClient('TEST', {
+            dispose: () => { },
+
+            onStateChange: () => { return { dispose: () => { } }; },
+
+            parseArgs: (path: string, purpose: ExternalRequestType): ClientUriDetails | null => {
+                if (purpose === ExternalRequestType.read_file)
+                    return {
+                        normalizedPath: () => '/DIR2/FILE',
+                    };
+                else if (path.endsWith('DIR'))
+                    return {
+                        normalizedPath: () => '/DIR',
+                    };
+                else if (path.endsWith('DIR2'))
+                    return {
+                        normalizedPath: () => '/DIR2',
+                    };
+                return null;
+            },
+
+            suspended: () => false,
+
+            uniqueId: () => Promise.resolve('SERVER'),
+        } as any as ExternalFilesClient);
+
+        const dir = await ext.handleRawMessage({ id: 4, op: 'read_directory', url: 'test:/TEST/DIR' });
+        assert.ok(dir);
+        assert.strictEqual(dir.id, 4);
+        assert.ok('error' in dir);
+        assert.strictEqual(dir?.error?.code, 0);
+
+        const dir2 = await ext.handleRawMessage({ id: 5, op: 'read_directory', url: 'test:/TEST/DIR2' });
+        assert.ok(dir2);
+        assert.strictEqual(dir2.id, 5);
+        assert.ok('data' in dir2);
+        assert.ok(dir2.data instanceof Object);
+        assert.deepStrictEqual(dir2.data.members, dir2Content);
+
+        const file = await ext.handleRawMessage({ id: 6, op: 'read_file', url: 'test:/TEST/DIR2/FILE' });
+        assert.ok(file);
+        assert.strictEqual(file.id, 6);
+        assert.ok('data' in file);
+        assert.strictEqual(file.data, fileContent);
+
+        ext.dispose();
+
+        assert.strictEqual(dirWritten, true);
+        assert.strictEqual(dir2Written, true);
+        assert.strictEqual(fileWritten, true);
     });
 });